+++ oshi-json/src/main/java/oshi/json/SystemInfo.java
@@ -18,16 +18,21 @@
  */
 package oshi.json;
 
+import java.util.Properties;
+
 import javax.json.Json;
 import javax.json.JsonBuilderFactory;
 import javax.json.JsonObject;
+import javax.json.JsonObjectBuilder;
 
 import oshi.json.hardware.HardwareAbstractionLayer;
 import oshi.json.hardware.impl.HardwareAbstractionLayerImpl;
+import oshi.json.json.AbstractOshiJsonObject;
 import oshi.json.json.NullAwareJsonObjectBuilder;
 import oshi.json.json.OshiJsonObject;
 import oshi.json.software.os.OperatingSystem;
 import oshi.json.software.os.impl.OperatingSystemImpl;
+import oshi.json.util.PropertiesUtil;
 
 /**
  * System information. This is the main entry point to Oshi. This object
@@ -37,7 +42,7 @@ import oshi.json.software.os.impl.OperatingSystemImpl;
  * 
  * @author dblock[at]dblock[dot]org
  */
-public class SystemInfo extends oshi.SystemInfo implements OshiJsonObject {
+public class SystemInfo extends AbstractOshiJsonObject implements OshiJsonObject {
 
     private static final long serialVersionUID = 1L;
 
@@ -83,9 +88,17 @@ public class SystemInfo extends oshi.SystemInfo implements OshiJsonObject {
      * {@inheritDoc}
      */
     @Override
-    public JsonObject toJSON() {
-        return NullAwareJsonObjectBuilder.wrap(jsonFactory.createObjectBuilder())
-                .add("platform", si.getCurrentPlatformEnum().toString())
-                .add("operatingSystem", getOperatingSystem().toJSON()).add("hardware", getHardware().toJSON()).build();
+    public JsonObject toJSON(Properties properties) {
+        JsonObjectBuilder json = NullAwareJsonObjectBuilder.wrap(jsonFactory.createObjectBuilder());
+        if (PropertiesUtil.getBoolean(properties, "platform")) {
+            json.add("platform", si.getCurrentPlatformEnum().toString());
+        }
+        if (PropertiesUtil.getBoolean(properties, "operatingSystem")) {
+            json.add("operatingSystem", getOperatingSystem().toJSON(properties));
+        }
+        if (PropertiesUtil.getBoolean(properties, "hardware")) {
+            json.add("hardware", getHardware().toJSON(properties));
+        }
+        return json.build();
     }
 }