+++ oshi-json/src/main/java/oshi/json/hardware/HWDiskStore.java
@@ -18,12 +18,16 @@
  */
 package oshi.json.hardware;
 
+import java.util.Properties;
+
 import javax.json.Json;
 import javax.json.JsonBuilderFactory;
 import javax.json.JsonObject;
+import javax.json.JsonObjectBuilder;
 
+import oshi.json.json.AbstractOshiJsonObject;
 import oshi.json.json.NullAwareJsonObjectBuilder;
-import oshi.json.json.OshiJsonObject;
+import oshi.json.util.PropertiesUtil;
 
 /**
  * A storage mechanism where data are recorded by various electronic, magnetic,
@@ -35,11 +39,13 @@ import oshi.json.json.OshiJsonObject;
  *
  * @author enrico[dot]bianchi[at]gmail[dot]com
  */
-public class HWDiskStore extends oshi.hardware.HWDiskStore implements OshiJsonObject {
+public class HWDiskStore extends AbstractOshiJsonObject {
 
     private static final long serialVersionUID = 1L;
 
     private JsonBuilderFactory jsonFactory = Json.createBuilderFactory(null);
+
+    private oshi.hardware.HWDiskStore hwDiskStore;
 
     /**
      * Create an object with all values
@@ -58,16 +64,123 @@ public class HWDiskStore extends oshi.hardware.HWDiskStore implements OshiJsonOb
      *            Number of writes to the disk
      */
     public HWDiskStore(String name, String model, String serial, long size, long reads, long writes) {
-        super(name, model, serial, size, reads, writes);
+        this.hwDiskStore = new oshi.hardware.HWDiskStore(name, model, serial, size, reads, writes);
+    }
+
+    /**
+     * @return the name
+     */
+    public String getName() {
+        return this.hwDiskStore.getName();
+    }
+
+    /**
+     * @param name
+     *            the name to set
+     */
+    public void setName(String name) {
+        this.hwDiskStore.setName(name);
+    }
+
+    /**
+     * @return the model
+     */
+    public String getModel() {
+        return this.hwDiskStore.getModel();
+    }
+
+    /**
+     * @param model
+     *            the model to set
+     */
+    public void setModel(String model) {
+        this.hwDiskStore.setModel(model);
+    }
+
+    /**
+     * @return the serial
+     */
+    public String getSerial() {
+        return this.hwDiskStore.getSerial();
+    }
+
+    /**
+     * @param serial
+     *            the serial to set
+     */
+    public void setSerial(String serial) {
+        this.hwDiskStore.setSerial(serial);
+    }
+
+    /**
+     * @return Get size of disk (in bytes)
+     */
+    public long getSize() {
+        return this.hwDiskStore.getSize();
+    }
+
+    /**
+     * @param size
+     *            Set size of disk (in bytes)
+     */
+    public void setSize(long size) {
+        this.hwDiskStore.setSize(size);
+    }
+
+    /**
+     * @return the reads
+     */
+    public long getReads() {
+        return this.hwDiskStore.getReads();
+    }
+
+    /**
+     * @param reads
+     *            the reads to set
+     */
+    public void setReads(long reads) {
+        this.hwDiskStore.setReads(reads);
+    }
+
+    /**
+     * @return the writes
+     */
+    public long getWrites() {
+        return this.hwDiskStore.getWrites();
+    }
+
+    /**
+     * @param writes
+     *            the writes to set
+     */
+    public void setWrites(long writes) {
+        this.hwDiskStore.setWrites(writes);
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public JsonObject toJSON() {
-        return NullAwareJsonObjectBuilder.wrap(jsonFactory.createObjectBuilder()).add("name", getName())
-                .add("model", getModel()).add("serial", getSerial()).add("size", getSize()).add("reads", getReads())
-                .add("writes", getWrites()).build();
+    public JsonObject toJSON(Properties properties) {
+        JsonObjectBuilder json = NullAwareJsonObjectBuilder.wrap(jsonFactory.createObjectBuilder());
+        if (PropertiesUtil.getBoolean(properties, "hardware.disks.name")) {
+            json.add("name", this.hwDiskStore.getName());
+        }
+        if (PropertiesUtil.getBoolean(properties, "hardware.disks.model")) {
+            json.add("model", this.hwDiskStore.getModel());
+        }
+        if (PropertiesUtil.getBoolean(properties, "hardware.disks.serial")) {
+            json.add("serial", this.hwDiskStore.getSerial());
+        }
+        if (PropertiesUtil.getBoolean(properties, "hardware.disks.size")) {
+            json.add("size", this.hwDiskStore.getSize());
+        }
+        if (PropertiesUtil.getBoolean(properties, "hardware.disks.reads")) {
+            json.add("reads", this.hwDiskStore.getReads());
+        }
+        if (PropertiesUtil.getBoolean(properties, "hardware.disks.writes")) {
+            json.add("writes", this.hwDiskStore.getWrites());
+        }
+        return json.build();
     }
 }