+++ oshi-json/src/main/java/oshi/json/software/os/impl/FileSystemImpl.java
@@ -18,16 +18,21 @@
  */
 package oshi.json.software.os.impl;
 
+import java.util.Properties;
+
 import javax.json.Json;
 import javax.json.JsonArrayBuilder;
 import javax.json.JsonBuilderFactory;
 import javax.json.JsonObject;
+import javax.json.JsonObjectBuilder;
 
+import oshi.json.json.AbstractOshiJsonObject;
 import oshi.json.json.NullAwareJsonObjectBuilder;
 import oshi.json.software.os.FileSystem;
 import oshi.json.software.os.OSFileStore;
+import oshi.json.util.PropertiesUtil;
 
-public class FileSystemImpl implements FileSystem {
+public class FileSystemImpl extends AbstractOshiJsonObject implements FileSystem {
 
     private static final long serialVersionUID = 1L;
 
@@ -74,13 +79,21 @@ public class FileSystemImpl implements FileSystem {
      * {@inheritDoc}
      */
     @Override
-    public JsonObject toJSON() {
-        JsonArrayBuilder fileStoreArrayBuilder = jsonFactory.createArrayBuilder();
-        for (OSFileStore fileStore : getFileStores()) {
-            fileStoreArrayBuilder.add(fileStore.toJSON());
+    public JsonObject toJSON(Properties properties) {
+        JsonObjectBuilder json = NullAwareJsonObjectBuilder.wrap(jsonFactory.createObjectBuilder());
+        if (PropertiesUtil.getBoolean(properties, "operatingSystem.fileSystem.fileStores")) {
+            JsonArrayBuilder fileStoreArrayBuilder = jsonFactory.createArrayBuilder();
+            for (OSFileStore fileStore : getFileStores()) {
+                fileStoreArrayBuilder.add(fileStore.toJSON(properties));
+            }
+            json.add("fileStores", fileStoreArrayBuilder.build());
         }
-        return NullAwareJsonObjectBuilder.wrap(jsonFactory.createObjectBuilder())
-                .add("fileStores", fileStoreArrayBuilder.build()).add("openFileDescriptors", getOpenFileDescriptors())
-                .add("maxFileDescriptors", getMaxFileDescriptors()).build();
+        if (PropertiesUtil.getBoolean(properties, "operatingSystem.fileSystem.openFileDescriptors")) {
+            json.add("openFileDescriptors", getOpenFileDescriptors());
+        }
+        if (PropertiesUtil.getBoolean(properties, "operatingSystem.fileSystem.maxFileDescriptors")) {
+            json.add("maxFileDescriptors", getMaxFileDescriptors());
+        }
+        return json.build();
     }
 }